backend/requirements.txt
aiogram==3.10.0
aiohttp==3.9.5
firebase-admin==6.5.0
google-auth==2.32.0
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
gcloud-aio-auth==5.3.2
gcloud-aio-storage==9.3.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
uvicorn==0.30.1
fastapi==0.111.1
